Add patch for ec config default in gitops pipeline
This is so future runs of hack/import-build-definitions do not
overwrite the customized ec-policy-configuration param default in
the gitops-pull-request pipeline.

Makes use of the pipeline patching support added in the previous
commit.

The commentary in the patch file gives some explanation why we want
that customization.

Note that running the hack/import-build-definition now brings in a
bunch of other changes to the tasks and pipelines. I have test this
patch by running that script, but I didn't check in the changes it
produced since I expect we don't want to bring them now so close to
a minor release.

Ref: https://issues.redhat.com/browse/EC-605